pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "huggingface-hub[cli]>=0.36.0",
    "punctuators==0.0.5",
    "pyannote-audio>=4.0.1",
    "torchaudio>=2.9.1",
    "transformers>=4.57.3",
]

[tool.uv.sources]
diarizers = { git = "https://github.com/huggingface/diarizers.git" }
